Split out failfast into smaller modules
* Moved `Gate` into its own module.
* Moved `test` into their own file.
* Reorganized `Gate:poll_ready` slightly for clarity.
* Added comments to `FailFast::poll_ready` state changes.
* Renamed `max_unavailable` to `timeout`.
